Fix Auth & Remove Forget for now
