+++ controllers/auth.js
@@ -2,6 +2,7 @@ const crypto = require('crypto');
 const  Course  = require('../models/classes/Course.js');// course not Course, Strange
 const  Student  = require('../models/classes/Student.js');
 const  Educator  = require('../models/classes/Educator.js');
+const { buildDataBeforeRender,createHash } = require('../middlewares/misc.js');
 
 // console.log('in pages.js');
 const { usersCollection, coursesCollection } = require('../models/config.js');
