+++ package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "connect-mongo": "^5.1.0",
     "dotenv": "^16.5.0",
+    "ejs": "^3.1.10",
     "express": "^5.1.0",
     "express-rate-limit": "^7.5.0",
     "express-session": "^1.18.1",
