+++ routes/routes.js
@@ -1,8 +1,8 @@
 const express= require('express')
 const router = express.Router();
 const controller = require('../controllers/pages')
-const controllerForget = require('../controllers/forget')
-const controllerAuth = require('../controllers/Auth')
+// const controllerForget = require('../controllers/forget')
+const controllerAuth = require('../controllers/auth')
 const controllerAdmin = require('../controllers/adminControllers')
 const rateLimit = require('express-rate-limit');
 console.log('in routes.js');
@@ -62,10 +62,10 @@ router.post('/login',checkUserLogin,controllerAuth.loginPost)
 router.get('/register',checkUserLogin,controllerAuth.registerGet)
 router.post('/register',checkUserLogin,controllerAuth.registerPost)
 router.get('/signout',controllerAuth.signout)
-router.get('/forget',controllerForget.forgetGet)
-router.post('/forget',controllerForget.forgetPost)
-router.get('/reset-password/:token',controllerForget.resetPasswordGet)
-router.post('/reset-password/:token',controllerForget.resetPasswordPost)
+// router.get('/forget',controllerForget.forgetGet)
+// router.post('/forget',controllerForget.forgetPost)
+// router.get('/reset-password/:token',controllerForget.resetPasswordGet)
+// router.post('/reset-password/:token',controllerForget.resetPasswordPost)
 router.get('/search',controller.search)
 router.get('/profile',controller.profile)
 router.get('/pricing',controller.pricing)
